Resolved issue #48
Fault was in the logic that allows users to temp partial shifts.
Only occurred when the entire shift was temped.
Resolved by checking to see if the entire shift was temped, but the form was used (this happens in the event that someone assigns their whole shift) and reverting to the whole shift logic if that were the case.